+++ domain/src/main/java/com/anytypeio/anytype/domain/dataview/interactor/DeleteDataViewViewer.kt
@@ -1,10 +1,10 @@
 package com.anytypeio.anytype.domain.dataview.interactor
 
+import com.anytypeio.anytype.core_models.Id
+import com.anytypeio.anytype.core_models.Payload
 import com.anytypeio.anytype.domain.base.BaseUseCase
 import com.anytypeio.anytype.domain.block.repo.BlockRepository
-import com.anytypeio.anytype.core_models.Id
 import com.anytypeio.anytype.domain.dataview.interactor.DeleteDataViewViewer.Params
-import com.anytypeio.anytype.core_models.Payload
 
 
 /**
@@ -28,7 +28,7 @@ class DeleteDataViewViewer(
      * @property [dataview] id of the data view containing a viewer, which we need to delete
      * @property [viewer] id of the viewer, which we need to delete
      */
-    class Params(
+    data class Params(
         val ctx: Id,
         val dataview: Id,
         val viewer: Id